Make /etc/make.tinderbox.private.conf the place to set BTI and PROXY settings.
Make bti in bashrc use the TINDERBOX_PROXY variable rather than BTI_PROXY,
and also set curl (with proxy) for fetch and resume.

Note: don't set the proxy in the environment or the isolation will fail;
unfortunately wget can be told to ignore proxies but _not_ to use them, on
the command line, which is why we went with curl.